+++ package.json
@@ -27,7 +27,6 @@
     "react-google-maps": "^9.4.5",
     "react-icons": "^4.2.0",
     "react-loadingg": "^1.7.2",
-    "react-map-gl": "^6.1.16",
     "react-modal-slideshow": "^1.0.5",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
