+++ src/App.js
@@ -4,7 +4,7 @@ import { useSelector } from "react-redux";
 import { Routes } from "./components/Route component/Routes";
 
 import HotCard from "./components/HotDet/HotCard";
-import CartLoad from "./components/HotDet/CardLoad";
+
 import Loading from "./components/material-ui-components/LoadingAnimation";
 // import { CheckoutPage } from "./components/Checkout-Page/CheckoutPage";
 
@@ -16,11 +16,6 @@ function App() {
   const login = useSelector((state) => state.login);
   return (
     <div>
-      <HotCard />
-      {login.isLoading && <Loading />}
-      <Routes />
-      <CartLoad />
-
       {/* {login.isLoading && <Loading />} */}
       {/* <Routes /> */}
